+++ Scripts/Structures/GameState/ResourceRequest.cs
@@ -6,11 +6,11 @@ namespace Structures.GameState;
 public class ResourceRequest
 {
     public Constructables.Building Building { get; }
-    public Dictionary<string, float> MissingResources { get; }
+    public Dictionary<string, int> MissingResources { get; }
     public int Priority { get; set; }
     public double Timestamp { get; }
 
-    public ResourceRequest(Constructables.Building building, Dictionary<string, float> missing, int priority, double timestamp)
+    public ResourceRequest(Constructables.Building building, Dictionary<string, int> missing, int priority, double timestamp)
     {
         Building = building;
         MissingResources = missing;
